Extract component MovieRow
